requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
PyMySQL==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.10.0
pydantic-settings==2.6.1
pillow==11.0.0
python-multipart==0.0.12
requests==2.32.3

